Updated solution again
