+++ problemsets/TestWorks/testwork-2-5/testwork-task01-requests.cpp
@@ -9,3 +9,58 @@ using namespace std;
 // Иначе следует добавить этот элемент в конец очереди.
 // Для каждого запроса -1 вывести убираемый элемент, либо -1, если очередь пуста.
 
+
+void handleInput(const int &n, int *requests)
+{
+	cout << "\nStarted testing module..." << endl;
+	cout << "\nGot array of requests:" << endl;
+	for (int i = 0; i < n; ++i)
+	{
+		cout << requests[i] << ", ";
+	}
+	
+	auto *myQueue = new Queue;
+	for (int i = 0; i < n; ++i)
+	{
+		if (requests[i] != -1)
+		{
+			cout << "Enqueued with value: " << requests[i] << endl;
+			myQueue->enqueue(requests[i]);
+		}
+		else
+		{
+			cout << "Dequeued value: " << myQueue->dequeue() << endl;
+		}
+	}
+	
+	cout << "At the end of the day we have: " << endl;
+	myQueue->printQueue();
+	delete myQueue;
+}
+
+
+void handleInput(const int &n, int *requests, Queue &myQueue)
+{
+	cout << "\nStarted testing module..." << endl;
+	cout << "\nGot array of requests:" << endl;
+	for (int i = 0; i < n; ++i)
+	{
+		cout << requests[i] << ", ";
+	}
+	
+	for (int i = 0; i < n; ++i)
+	{
+		if (requests[i] != -1)
+		{
+			cout << "Enqueued with value: " << requests[i] << endl;
+			myQueue.enqueue(requests[i]);
+		}
+		else
+		{
+			cout << "Dequeued value: " << myQueue.dequeue() << endl;
+		}
+	}
+	
+	cout << "At the end of the day we have: " << endl;
+	myQueue.printQueue();
+}
